+++ src/keys.rs
@@ -6,7 +6,7 @@
 
 use alloc::vec::Vec;
 use core::fmt;
-use core2::io::{self, Read, Write};
+use corez::io::{self, Read, Write};
 
 use super::{
     address::PaymentAddress,
@@ -346,7 +346,7 @@ impl ProofGenerationKey {
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
 pub struct NullifierDerivingKey(pub jubjub::SubgroupPoint);
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, PartialEq, Eq)]
 pub struct ViewingKey {
     pub ak: SpendValidatingKey,
     pub nk: NullifierDerivingKey,
@@ -367,7 +367,7 @@ impl ViewingKey {
 }
 
 /// A Sapling key that provides the capability to view incoming and outgoing transactions.
-#[derive(Debug)]
+#[derive(Debug, PartialEq, Eq)]
 pub struct FullViewingKey {
     pub vk: ViewingKey,
     pub ovk: OutgoingViewingKey,
@@ -447,7 +447,7 @@ impl FullViewingKey {
     }
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, PartialEq, Eq)]
 pub struct SaplingIvk(pub jubjub::Fr);
 
 impl SaplingIvk {