+++ CHANGELOG.md
@@ -7,9 +7,25 @@ and this library adheres to Rust's notion of
 
 ## [Unreleased]
 
+## [0.7.0] - 2026-04-21
+
 ### Changed
+- MSRV is now 1.85.1
+- Migrated from the yanked `core2` crate to `corez` for no_std IO support.
 - `sapling_crypto::pczt::Bundle::extract` now takes its `self` argument by
   reference instead of by value.
+
+## [0.6.2] - 2026-03-31
+
+- `impl {PartiaEq, Eq} for sapling_crypto::keys::SaplingIvk`
+- `impl {PartiaEq, Eq} for sapling_crypto::zip32::IncomingViewingKey`
+
+## [0.6.1] - 2026-03-26
+
+### Added
+- `impl {PartiaEq, Eq} for sapling_crypto::keys::ViewingKey`
+- `impl {PartiaEq, Eq} for sapling_crypto::keys::FullViewingKey`
+- `impl {PartiaEq, Eq} for sapling_crypto::zip32::DiversifiableFullViewingKey`
 
 ## [0.6.0] - 2025-12-05